+++ doctr/models/recognition/sar/pytorch.py
@@ -241,7 +241,7 @@ class SAR(nn.Module, RecognitionModel):
         mask_2d = torch.arange(input_len, device=model_output.device)[None, :] < seq_len[:, None]
         cce[mask_2d] = 0
 
-        ce_loss = cce.sum(1) / seq_len.to(dtype=torch.float32)
+        ce_loss = cce.sum(1) / seq_len.to(dtype=model_output.dtype)
         return ce_loss.unsqueeze(1)
